+++ apis/QryAsset.go
@@ -7,6 +7,11 @@ import (
 	"github.com/gorilla/mux"
 )
 
+/*
+	ServeAssetWhole
+	- API Query processing with Vars
+	- TODO: Change it to r.URL.Query - this enables optional existance query
+*/
 func ServeAssetWhole(d IDPAsset, w http.ResponseWriter, r *http.Request) {
 	params := mux.Vars(r)
 	qry := procAssetParam(params)
@@ -18,6 +23,10 @@ func ServeAssetWhole(d IDPAsset, w http.ResponseWriter, r *http.Request) {
 	w.Write(packet)
 }
 
+/*
+	procAssetParam
+	- Store queries in ReqIDPAsset
+*/
 func procAssetParam(m map[string]string) ReqIDPAsset {
 	// strat in ["All", "Core", "Value-added", "Opportunistic"]
 	q := ReqIDPAsset{
@@ -26,6 +35,10 @@ func procAssetParam(m map[string]string) ReqIDPAsset {
 	return q
 }
 
+/*
+	procAssetQry
+	- find row inside data that matches condition
+*/
 func procAssetQry(rq ReqIDPAsset, d IDPAsset) []assets {
 	// reads asset query from ReqIDPAsset struct
 	var sendPacket = []assets{}